new common completer class
