+++ ossdbtoolsservice/language/completion/pgcompleter.py
@@ -26,6 +26,7 @@ from .packages.parseutils.utils import last_word
 from .packages.parseutils.tables import TableReference
 from .packages.pgliterals.main import get_literals
 from .packages.prioritization import PrevalenceCounter
+from .completer import MyCompleter, _Candidate, Match
 # {{ PGToolsService EDIT }}
 # from .config import load_config, config_location
 
@@ -34,8 +35,6 @@ from .packages.prioritization import PrevalenceCounter
 # {{ PGToolsService EDIT }}
 # NamedQueries.instance = NamedQueries.from_config(
 #     load_config(config_location() + 'config'))
-
-Match = namedtuple('Match', ['completion', 'priority'])
 
 _SchemaObject = namedtuple('SchemaObject', 'name schema meta schema_name')
 
@@ -46,11 +45,6 @@ def SchemaObject(name, schema=None, meta=None, schema_name=None):
     need it based on _maybe_schema logic. schema_name always holds a value.
     """
     return _SchemaObject(name, schema, meta, schema_name)
-
-
-_Candidate = namedtuple(
-    'Candidate', 'completion prio meta synonyms prio2 display schema'
-)
 
 
 def Candidate(
@@ -79,7 +73,7 @@ def generate_alias(tbl):
                    [letter for letter, prev in zip(tbl, '_' + tbl) if prev == '_' and letter != '_'])
 
 
-class PGCompleter(Completer):
+class PGCompleter(Completer, MyCompleter):
     # keywords_tree: A dict mapping keywords to well known following keywords.
     # e.g. 'CREATE': ['TABLE', 'USER', ...],
     keywords_tree = get_literals('keywords', type_=dict)
@@ -89,10 +83,9 @@ class PGCompleter(Completer):
     reserved_words = set(get_literals('reserved'))
 
     def __init__(self, smart_completion=True, logger=None, settings=None):
-        super(PGCompleter, self).__init__()
+        super(PGCompleter, self).__init__(PGCompletion)
         self.smart_completion = smart_completion
         self.logger: Logger = logger
-        self.prioritizer = PrevalenceCounter()
         settings = settings or {}
         self.signature_arg_style = settings.get(
             'signature_arg_style', '{arg_name} {arg_type}'
@@ -129,7 +122,6 @@ class PGCompleter(Completer):
         self.dbmetadata = {'tables': {}, 'views': {}, 'functions': {},
                            'datatypes': {}}
         self.search_path = []
-        self.casing = {}
 
         self.all_completions = set(self.keywords + self.functions)
 
@@ -150,13 +142,6 @@ class PGCompleter(Completer):
 
     def escape_schema(self, name):
         return "'{}'".format(self.unescape_name(name))
-
-    def unescape_name(self, name):
-        """ Unquote a string."""
-        if name and name[0] == '"' and name[-1] == '"':
-            name = name[1:-1]
-
-        return name
 
     def escaped_names(self, names):
         return [self.escape_name(name) for name in names]
@@ -327,132 +312,6 @@ class PGCompleter(Completer):
                            'datatypes': {}}
         self.all_completions = set(self.keywords + self.functions)
 
-    def find_matches(self, text, collection, mode='fuzzy', meta=None):
-        """Find completion matches for the given text.
-
-        Given the user's input text and a collection of available
-        completions, find completions matching the last word of the
-        text.
-
-        `collection` can be either a list of strings or a list of Candidate
-        namedtuples.
-        `mode` can be either 'fuzzy', or 'strict'
-            'fuzzy': fuzzy matching, ties broken by name prevalance
-            `keyword`: start only matching, ties broken by keyword prevalance
-
-        yields prompt_toolkit Completion instances for any matches found
-        in the collection of available completions.
-
-        """
-        if not collection:
-            return []
-        prio_order = [
-            'keyword', 'function', 'view', 'table', 'datatype', 'database',
-            'schema', 'column', 'table alias', 'join', 'name join', 'fk join'
-        ]
-        type_priority = prio_order.index(meta) if meta in prio_order else -1
-        text = last_word(text, include='most_punctuations').lower()
-        text_len = len(text)
-
-        if text and text[0] == '"':
-            # text starts with double quote; user is manually escaping a name
-            # Match on everything that follows the double-quote. Note that
-            # text_len is calculated before removing the quote, so the
-            # Completion.position value is correct
-            text = text[1:]
-
-        if mode == 'fuzzy':
-            fuzzy = True
-            priority_func = self.prioritizer.name_count
-        else:
-            fuzzy = False
-            priority_func = self.prioritizer.keyword_count
-
-        # Construct a `_match` function for either fuzzy or non-fuzzy matching
-        # The match function returns a 2-tuple used for sorting the matches,
-        # or None if the item doesn't match
-        # Note: higher priority values mean more important, so use negative
-        # signs to flip the direction of the tuple
-        if fuzzy:
-            regex = '.*?'.join(map(re.escape, text))
-            pat = re.compile('(%s)' % regex)
-
-            def _match(item):
-                if item.lower()[:len(text) + 1] in (text, text + ' '):
-                    # Exact match of first word in suggestion
-                    # This is to get exact alias matches to the top
-                    # E.g. for input `e`, 'Entries E' should be on top
-                    # (before e.g. `EndUsers EU`)
-                    return float('Infinity'), -1
-                r = pat.search(self.unescape_name(item.lower()))
-                if r:
-                    return -len(r.group()), -r.start()
-        else:
-            match_end_limit = len(text)
-
-            def _match(item):
-                match_point = item.lower().find(text, 0, match_end_limit)
-                if match_point >= 0:
-                    # Use negative infinity to force keywords to sort after all
-                    # fuzzy matches
-                    return -float('Infinity'), -match_point
-
-        matches = []
-        for cand in collection:
-            if isinstance(cand, _Candidate):
-                item, prio, display_meta, synonyms, prio2, display, schema = cand
-                if display_meta is None:
-                    display_meta = meta
-                syn_matches = (_match(x) for x in synonyms)
-                # Nones need to be removed to avoid max() crashing in Python 3
-                syn_matches = [m for m in syn_matches if m]
-                sort_key = max(syn_matches) if syn_matches else None
-            else:
-                item, display_meta, prio, prio2, display, schema = cand, meta, 0, 0, cand, cand
-                sort_key = _match(cand)
-
-            if sort_key:
-                if display_meta and len(display_meta) > 50:
-                    # Truncate meta-text to 50 characters, if necessary
-                    display_meta = display_meta[:47] + u'...'
-
-                # Lexical order of items in the collection, used for
-                # tiebreaking items with the same match group length and start
-                # position. Since we use *higher* priority to mean "more
-                # important," we use -ord(c) to prioritize "aa" > "ab" and end
-                # with 1 to prioritize shorter strings (ie "user" > "users").
-                # We first do a case-insensitive sort and then a
-                # case-sensitive one as a tie breaker.
-                # We also use the unescape_name to make sure quoted names have
-                # the same priority as unquoted names.
-                lexical_priority = (tuple(0 if c in (' _') else -ord(c)
-                                          for c in self.unescape_name(item.lower())) + (1,)
-                                    + tuple(c for c in item))
-
-                item = self.case(item)
-                display = self.case(display)
-                priority = (
-                    sort_key, type_priority, prio, priority_func(item),
-                    prio2, lexical_priority
-                )
-
-                extend_completion = PGCompletion(
-                    text=item,
-                    start_position=-text_len,
-                    display_meta=display_meta,
-                    display=display,
-                    schema=schema)
-
-                matches.append(
-                    Match(
-                        completion=extend_completion,
-                        priority=priority
-                    )
-                )
-        return matches
-
-    def case(self, word):
-        return self.casing.get(word, word)
 
     def get_completions(self, document, complete_event, smart_completion=None) -> List[Completion]:
         word_before_cursor = document.get_word_before_cursor(WORD=True)
